commander: fix race condition for mission result
This fixes the case where a second auto-takeoff fails because
_mission_result.finished is not updated in time so when we check it if
the takeoff is finished, we still look at the result of the previous
takeoff.

By comparing the timestamp of the _mission_result against the timestamp
of the internal state where the takeoff state is set, we can check the
validity of _mission_result.